+++ application/controllers/Sales.php
@@ -1,6 +1,14 @@
 <?php
 
-//defined('BASEPATH') OR exit('No direct script access allowed');
+defined('BASEPATH') OR exit('No direct script access allowed');
+
+/**
+ * Sales page allows any user to order products from the store. 
+ * The users can check the orders they already ordered, and check details as well.
+ * When start a new order, the page will show all the products that are selling in 
+ * the store, and users can order any products that are avaliable in the stock.
+ * When cancel the order, return the orders page.
+ **/
 
 class Sales extends Application
 {
@@ -8,6 +16,7 @@ class Sales extends Application
     function __construct()
     {
         parent::__construct();
+        $this->data['pagetitle'] = 'Sales';
     }
 
     /**
@@ -18,12 +27,6 @@ class Sales extends Application
         if ($this->session->has_userdata('orders'))
             $this->keep_shopping();
         else $this->summarize();
-
-        /*$stuff = file_get_contents('../data/receipt.md');
-        $this->data['order'] = $this->parsedown->parse($stuff);
-        $this->data['pagebody'] = 'sales_view';
-		$this->data['items'] = $this->recipes->all();
-		$this->render('template');*/
     }
 
     public function summarize() 
@@ -68,7 +71,7 @@ class Sales extends Application
         $this->data['content'] = '';
         $this->data['items'] = '';
         foreach($this->recipes->all() as $menuitem) {
-           $this->data['items'] .= $this->parser->parse('menuitem-shop',$menuitem,true);
+           $this->data['items'] .= $this->parser->parse('sales-order_view',$menuitem,true);
         }
         $this->render('sales_view'); 
 	}
@@ -76,7 +79,8 @@ class Sales extends Application
     public function add($what) 
     {
         $orders = new Orders($this->session->userdata('orders'));
-        $orders->additem($what);
+        $orders->additem($what); 
+        $orders->updateRecipes($what);     
         $this->keep_shopping();
         $this->session->set_userdata('orders',(array)$orders);
         redirect('/Sales');
@@ -92,24 +96,33 @@ class Sales extends Application
         $this->index();
     }
 
-    public function checkout() 
+    public function checkout()
     {
         $orders = new Orders($this->session->userdata('orders'));
+        
         // ignore invalid requests
         /*if (! $orders->validate())
             redirect('/Sales');*/
-
         $orders->save();
         $this->session->unset_userdata('orders');
+
+        // Calculate store's running total
+        $this->load->helper('file');
+        $currentTotal = file_get_contents('../data/money.txt');
+        $newRunningTotal = $orders->total() + $currentTotal;
+        if ( ! write_file('../data/money.txt', $newRunningTotal))
+        {
+            echo 'Unable to write the file';
+        }
         redirect('/Sales');
+
     }
 
     public function examine($which)
-     {
+    {
         $orders = new Orders ('../data/order' . $which . '.xml');
-        $stuff = $orders->receipt();
+        $stuff = $orders->receipt($which);
         $this->data['content'] = $this->parsedown->parse($stuff);
         $this->render();
     }
-
 }